requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.7
google-cloud-vision==3.4.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3